Rearrange code of wayland frontend
